[TASK] Introduce GitHub action workflow as CI check
This change introduces GitHub action workflows to
integrate automatic CI checks for pushes and pull
requests.

This change adds a general php-linting for supported
php-versions and executes unit- and integration tests
against laravel 9.x and 10.x.

This helps contributors a lot, and also helps the
maintainer(s) for pull-requests, that tests are
green. Maybe this helps to minimze manual work
for pull-request reviews and increases merging
working pull-requests.

Resolves #110